Replaced IAM names with name_prefix (#18)
* Replaced IAM names with name_prefix

This replaces all the IAM Name values, with `name_prefix` thus ensuring
unique names. I'm not sure I *don't* think this is awful. But - meh -
it'll allow this module to be used more easily across regions.

* This should now include the correct offsets

This should now inlcude the correct charecter offsets for a 26 charecter
prefix, based on the IAM resource limits found here
https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_iam-limits.html

* This compiles correctly

Not sure why, as 32+26 = 58 not 64... maybe I'm reading the doc wrong?